// frontend/src/lib/api/types/AgentCreateRequest.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


/**
 * @description Minimal request schema for creating a new agent via our API.
*/
export type AgentCreateRequest = {
    /**
     * @description The system prompt that defines the agent\'s behavior and personality
     * @minLength 1
     * @type string
    */
    prompt: string;
    /**
     * @description Optional name for the agent (for internal reference)
    */
    agent_name?: (string | null);
};